+++ src/Exception/CoordinateSystemException.php
@@ -11,7 +11,16 @@ use Brick\Geo\Geometry;
  */
 class CoordinateSystemException extends GeometryException
 {
-    public static function sridMix(Geometry $reference, Geometry $culprit) : CoordinateSystemException
+    public static function sridMix(int $srid1, int $srid2) : CoordinateSystemException
+    {
+        return new CoordinateSystemException(sprintf(
+            'SRID mix: cannot mix SRID %d with SRID %d.',
+            $srid1,
+            $srid2
+        ));
+    }
+
+    public static function sridCompositionMix(Geometry $reference, Geometry $culprit) : CoordinateSystemException
     {
         return new CoordinateSystemException(sprintf(
             'SRID mix: %s with SRID %d cannot contain %s with SRID %d.',